Pick the first server according to sort_order.
The handling of sort_order does not work correct for the sever list on
the login page. The problem is that connection.serverId, the selected
element, is not always a reference value to a server in the list, since
it can have been read from the localStorage. Angular however compares
ng-model by reference, so it doesn't work perfectly there. It is also not
possible to sort the ng-options since $scope.servers is a object itself.